convert logging to new API lg2 from log
This commit convert the logging function call to lg2.

In the old unit test, there are mock up interfaces of gpio and hwmonio.
The old log API is triggered when realworld failure events happen, but not by the dummy value of the mock up objects.
The phosphor-hwmon log would not be seen from journalctl.

lg2 unittest is added to test by directly calling BuildGpioHandle code
not through the mock up interface.

Tested: With build and excute the new lg2 unit test, the logging message can be seen from
journalctl.

Change-Id: I9fbab6c484ae3ce99e1de8eb1a10a78713d13f85